+++ docs/changelog.md
@@ -1,12 +1,14 @@
-# Changelog for Tabular for Python
+# Changelog for Tabulator for Python
 
-## pytabular v0.2.0
+## pytabulator v0.2.0
 
 * Add docs
 * Rename `TabulatorOptions` to `TableOptions`
 * Use Pydantic for `TableOptions`
 * Add more `input` events
+* Add detailed example
+* Add utility function to create columns configuration from data frame
 
-## pytabular v0.1.0 (2024-01-31, Pre-release)
+## pytabulator v0.1.0 (2024-01-31, Pre-release)
 
 Initial release